1.0.0

+++ deps.ts
@@ -6,6 +6,6 @@ export {
 export { TextEncoding } from "https://raw.githubusercontent.com/i-xi-dev/text-encoding.es/4.2.0/mod.ts";
 export {
   SafeInteger,
-  Uint16,
+  Uint32,
   Uint8,
 } from "https://raw.githubusercontent.com/i-xi-dev/number.es/6.0.0/mod.ts";
